Add print part for modules.py
